Started the pattern matching version of Eliza
I need to add the input parser next.  Since we're dealing with strings,
instead of list of atoms, it will need to balance parentheses and double
quotes. One of the exercises is to handle punctuation correctly.  I'm
going to strip it out since our list of patterns doesn't contain any.